Massive update to bring the project up-to-date and include some more recent technologies.
Still using Hapi, but extended the base to be more flexible, offer more examples, and to incliude basic authentication, a basic website layout (using Bootstrap), templating (via Nunjucks), and cleaner, more readable/configurable settings and code.

There is still quite a bit that needs to be done, and the code needs a bit of cleaning up, but this is a good start.